requirements.txt
pyaudio
SpeechRecognition
standard-aifc
pyttsx3
google-generativeai
python-dotenv